requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
coverage==7.6.9
cryptography==41.0.0
cycler==0.12.1
fastapi==0.115.4
fonttools==4.55.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
kiwisolver==1.4.7
Mako==1.3.6
MarkupSafe==3.0.2
matplotlib==3.9.3
numpy==2.2.0
packaging==24.2
pdfkit==1.0.0
pillow==11.0.0
pluggy==1.5.0
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
pytest==8.3.3
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
tqdm==4.67.0
typing_extensions==4.12.2
uvicorn==0.32.0
